Add scorekeeping and game(), fixed problems
Function for looping the game five times added. Then realised that the 'random' number returned is always the same. Quick google search revealed, that the function was nested too dep so fixed that. Also realised I had forgotten to add conditions for the player choosing scissors, so added those. Finally, score adjustment was added and the statements adjusted.